Работа с файлами. Пункты 2,3,4//ДЗ

+++ HomeWorks/Bot/Infrastructure/DataAccess/FileToDoRepository.cs
@@ -5,6 +5,8 @@ namespace Bot;
 public class FileToDoRepository : IToDoRepository
 {
     private string _baseDirecory;
+
+    private string _indexFile;
     
     public FileToDoRepository(string baseFolder)
     {
@@ -13,35 +15,79 @@ public class FileToDoRepository : IToDoRepository
             Directory.CreateDirectory(baseFolder);
         }
         _baseDirecory = baseFolder;
-    }
 
-    private async Task<IEnumerable<ToDoItem>> GetAll()
+        _indexFile = Path.Combine(_baseDirecory, "index.json");
+            
+        if (!File.Exists(_indexFile))
+        {
+            CreateFileIndex();
+        }
+    }
+    private async Task WriteFileIndex(List<Index>  indexes)
     {
-        return Directory.EnumerateFiles(_baseDirecory)
-            .ToList()
-            .Select((file) =>
-            {
-                try
-                {
-                    return JsonSerializer.Deserialize<ToDoItem>(File.ReadAllText(file));
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Error proccesing file {file}:\n{ex.Message}");
-                    return null;
-                }
-            })
-            .Where(item => item != null);
+        await File.WriteAllTextAsync(_indexFile, JsonSerializer.Serialize<List<Index>>(indexes, new JsonSerializerOptions { WriteIndented = true }));
     }
-
+    private async Task<List<Index>> ReadFileIndex()
+    {
+        if (!File.Exists(_indexFile))
+        {
+            await CreateFileIndex();
+        }        
+        return JsonSerializer.Deserialize<List<Index>>(File.ReadAllText(_indexFile));
+    }
+    private async Task CreateFileIndex()
+    {
+        await WriteFileIndex(
+            Directory.GetFiles(_baseDirecory, searchOption: SearchOption.AllDirectories, searchPattern: "*.json")
+                .Select((file) =>
+                {
+                    try
+                    {
+                        return JsonSerializer.Deserialize<ToDoItem>(File.ReadAllText(file));
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"Error proccesing file {file}:\n{ex.Message}");
+                        return null;
+                    }
+                })
+                .Where(item => item != null)
+                .Select(item => new Index(item.ToDoUser.UserId, item.Id)).ToList());
+    }
     public async Task<ToDoItem?> GetByGuid(Guid id)
     {
-        return (await GetAll()).FirstOrDefault(item => item.Id == id);
+        return (await Find(
+            (await ReadFileIndex()).FirstOrDefault(record => record.ItemId == id)?.UserId ?? Guid.Empty,
+            item => item.Id == id)).FirstOrDefault();
     }
 
     public async Task<IReadOnlyList<ToDoItem>> GetAllByUserId(Guid userId)
     {
-        return (await GetAll()).Where(item => item.ToDoUser.UserId == userId).ToList().AsReadOnly();
+        var userFolder = Path.Combine(_baseDirecory, userId.ToString());
+        if (!Directory.Exists(userFolder))
+        {
+            return new List<ToDoItem>().AsReadOnly();
+        }
+        else
+        {
+            return Directory.EnumerateFiles(userFolder)
+                .Where(file => Path.GetExtension(file) == ".json")
+                .Select((file) =>
+                {
+                    try
+                    {
+                        return JsonSerializer.Deserialize<ToDoItem>(File.ReadAllText(file));
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"Error proccesing file {file}:\n{ex.Message}");
+                        return null;
+                    }
+                })
+                .Where(item => item != null && item.ToDoUser.UserId == userId)
+                .ToList()
+                .AsReadOnly();
+        }
     }
 
     public async Task<IReadOnlyList<ToDoItem>> GetActiveByUserId(Guid userId)
@@ -51,22 +97,36 @@ public class FileToDoRepository : IToDoRepository
 
     public async Task Add(ToDoItem item)
     {
-        var filePath = Path.Combine(_baseDirecory, item.Id.ToString());
-        await File.WriteAllTextAsync(filePath, JsonSerializer.Serialize(item));   
+        var index = await ReadFileIndex();
+        var userFolder = Path.Combine(_baseDirecory, item.ToDoUser.UserId.ToString());
+        if (!Directory.Exists(userFolder))
+        {
+            Directory.CreateDirectory(userFolder);
+        }
+        var filePath = Path.Combine(userFolder, $"{item.Id}.json");
+        await File.WriteAllTextAsync(filePath, JsonSerializer.Serialize(item, new JsonSerializerOptions { WriteIndented = true }));
+        index.Add(new Index(item.ToDoUser.UserId, item.Id));
+        await WriteFileIndex(index);
     }
 
     public async Task Update(ToDoItem item)
     {
-        var filePath = Path.Combine(_baseDirecory, item.Id.ToString());
-        await File.WriteAllTextAsync(filePath, JsonSerializer.Serialize(item));   
+        var filePath = Path.Combine(_baseDirecory, item.ToDoUser.UserId.ToString(), $"{item.Id.ToString()}.json");
+        await File.WriteAllTextAsync(filePath, JsonSerializer.Serialize(item, new JsonSerializerOptions { WriteIndented = true }));   
     }
 
     public async Task Delete(Guid id)
     {
-        var filePath = Path.Combine(_baseDirecory, id.ToString());
-        if (File.Exists(filePath))
+        var indexes = (await ReadFileIndex());
+        var index = indexes.FirstOrDefault(record => record.ItemId == id);
+        if (index != null)
         {
-            File.Delete(filePath);
+            indexes.Remove(index);
+            var filePath = Path.Combine(_baseDirecory, index.UserId.ToString(), $"{id.ToString()}.json");
+            if (File.Exists(filePath))
+            {
+                File.Delete(filePath);
+            }
         }
     }
